moved error to domain package
