+++ app/core/domain/error/errors.go
@@ -1,4 +1,4 @@
-package inbound
+package error
 
 import "fmt"
