+++ app/core/domain/error/errors_test.go
@@ -1,4 +1,4 @@
-package inbound
+package error
 
 import (
 	"testing"
